backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
influxdb-client==1.45.0
python-dotenv==1.0.1
mysqlclient==2.2.4